requirements.txt
fastapi
uvicorn
requests
pydantic
streamlit
python-dotenv
openai
tiktoken